- order of operations for array access slightly optimized - fixed error in how float constants were compiled in oztree.c (there was an error with casting, I think code had been copy-pasted from the int const function... forgot to change int* to float*)

+++ oztree.c
@@ -961,7 +961,7 @@ gen_real_const(OzProgram *p, int reg, float val) {
     int *preg = checked_malloc(sizeof(int));
     *preg = reg;
 
-    int *pval = checked_malloc(sizeof(float));
+    float *pval = checked_malloc(sizeof(float));
     *pval = val;
 
     OzOp *op = new_op(p);
